Moved router to it's own file and calling it from index. index now initializes a navbar object which will stay in memory rather than constantly recreating it with each refresh
