+++ _main_project/static/js/helpers/helpers.js
@@ -1,3 +1,6 @@
+import { router } from './router.js';
+
+// Checks the login status of the user.
 export const loginCheck = async (return_data = false) => {
   try {
     const response = await fetch('/login_check/', {
@@ -24,3 +27,22 @@ export const loginCheck = async (return_data = false) => {
     return false;
   }
 };
+
+// Create a regex to replace the path with something.
+export const pathToRegex = path => new RegExp('^' + path.replace(/\//g, "\\/").replace(/:\w+/g, '(.+)') + '$');
+
+// Finds the key and the values for our params
+export const getParams = match => {
+  const values = match.result.slice(1);
+  const keys = Array.from(match.route.path.matchAll(/:(\w+)/g)).map(result => result[1]);
+
+  return Object.fromEntries(keys.map((key, i) => {
+    return [key, values[i]];
+  }));
+}
+
+// Allows us to use the history API to navigate to different routes
+export const navigateTo = url => {
+  history.pushState(null, null, url);
+  router();
+}
